+++ tests/MoBi.Tests/Helpers/DomainHelperForSpecs.cs
@@ -1,13 +1,11 @@
 ﻿using System;
 using System.IO;
-using MoBi.Core.Domain.Builder;
 using MoBi.Core.Domain.Model;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Data;
 using OSPSuite.Core.Domain.Services;
 using OSPSuite.Core.Domain.UnitSystem;
-using OSPSuite.Utility.Container;
 
 namespace MoBi.Helpers
 {
@@ -30,19 +28,19 @@ namespace MoBi.Helpers
          return Path.Combine(dataFolder, fileName);
       }
 
-      public static IDimension AmountDimension { get; } = new Dimension(new BaseDimensionRepresentation {AmountExponent = 1}, Constants.Dimension.MOLAR_AMOUNT, "µmol");
+      public static IDimension AmountDimension { get; } = new Dimension(new BaseDimensionRepresentation { AmountExponent = 1 }, Constants.Dimension.MOLAR_AMOUNT, "µmol");
 
       public static ParameterStartValue ParameterStartValue { get; } = new ParameterStartValue { Name = "_name", Value = 1.0, Formula = null, Dimension = AmountDimension };
 
-      public static IDimension ConcentrationDimension { get; } = new Dimension(new BaseDimensionRepresentation {LengthExponent = -3, MassExponent = 1, TimeExponent = -1}, Constants.Dimension.MOLAR_CONCENTRATION, "µmol/l");
+      public static IDimension ConcentrationDimension { get; } = new Dimension(new BaseDimensionRepresentation { LengthExponent = -3, MassExponent = 1, TimeExponent = -1 }, Constants.Dimension.MOLAR_CONCENTRATION, "µmol/l");
 
-      public static IDimension FractionDimension { get; } = new Dimension(new BaseDimensionRepresentation() , Constants.Dimension.FRACTION, "");
+      public static IDimension FractionDimension { get; } = new Dimension(new BaseDimensionRepresentation(), Constants.Dimension.FRACTION, "");
 
-      public static IDimension ConcentrationPerTimeDimension { get; } = new Dimension(new BaseDimensionRepresentation {LengthExponent = -3, AmountExponent = 1, TimeExponent = -1}, Constants.Dimension.MOLAR_CONCENTRATION_PER_TIME, "µmol/l/min");
+      public static IDimension ConcentrationPerTimeDimension { get; } = new Dimension(new BaseDimensionRepresentation { LengthExponent = -3, AmountExponent = 1, TimeExponent = -1 }, Constants.Dimension.MOLAR_CONCENTRATION_PER_TIME, "µmol/l/min");
 
-      public static IDimension AmountPerTimeDimension { get; } = new Dimension(new BaseDimensionRepresentation {AmountExponent = 1, TimeExponent = -1}, Constants.Dimension.AMOUNT_PER_TIME, "µmol/min");
+      public static IDimension AmountPerTimeDimension { get; } = new Dimension(new BaseDimensionRepresentation { AmountExponent = 1, TimeExponent = -1 }, Constants.Dimension.AMOUNT_PER_TIME, "µmol/min");
 
-      public static IDimension TimeDimension { get; } = new Dimension(new BaseDimensionRepresentation {TimeExponent = 1}, Constants.Dimension.TIME, "min");
+      public static IDimension TimeDimension { get; } = new Dimension(new BaseDimensionRepresentation { TimeExponent = 1 }, Constants.Dimension.TIME, "min");
 
       public static DataRepository ObservedData(string id = "TestData", IDimension timeDimension = null, IDimension concentrationDimension = null, string obsDataColumnName = null)
       {
@@ -71,15 +69,7 @@ namespace MoBi.Helpers
 
       public static IMoBiProject NewProject()
       {
-         return new MoBiProject(new SimulationSettingsFactoryForSpecs());
-      }
-   }
-
-   public class SimulationSettingsFactoryForSpecs : ISimulationSettingsFactory
-   {
-      public SimulationSettings CreateDefault()
-      {
-         return new SimulationSettings();
+         return new MoBiProject { SimulationSettings = new SimulationSettings() };
       }
    }
